Move authentication cache to tenant space
